+++ resources/views/livewire/item/item-edit.blade.php
@@ -1,8 +1,17 @@
 <div>
     <x-layouts.header :name="$item->name . ' (' . $item->code . ')'"/>
     <x-layouts.actions>
-        <flux:button wire:click="save">Сохранить</flux:button>
-        <flux:button variant="danger" wire:click="destroy">Удалить</flux:button>
+        @if($this->user()->can('update-items'))
+            <flux:button wire:click="save">Сохранить</flux:button>
+        @endif
+        @if($this->user()->can('delete-items'))
+            <flux:button
+                    variant="danger"
+                    wire:click="destroy"
+                    wire:confirm="Вы действительно хотите удалить этот товар?"
+            >Удалить
+            </flux:button>
+        @endif
     </x-layouts.actions>
     <x-layouts.main-container>
         <flux:tab.group>
@@ -24,7 +33,8 @@
                                 <flux:input wire:model="form.ms_uuid" label="МС UUID"/>
                                 <flux:input wire:model="form.article" label="Артикул поставщик" required/>
                                 <flux:input wire:model="form.brand" label="Бренд поставщик"/>
-                                <flux:input wire:model="form.multiplicity" label="Кратность отгрузки" type="number" required/>
+                                <flux:input wire:model="form.multiplicity" label="Кратность отгрузки" type="number"
+                                            required/>
                                 <flux:input wire:model="form.buy_price_reserve" label="Резервная закупочная цена"
                                             type="number"/>
                                 <flux:select variant="combobox" placeholder="Выберите поставщика..."
@@ -44,28 +54,41 @@
                                             <div class="flex justify-between items-center">
                                                 <flux:switch wire:model="form.attributes.{{$mainAttribute->id}}"
                                                              :label="$mainAttribute->name"/>
-                                                <flux:button wire:click="deleteAttribute({{$mainAttribute}})" variant="danger"
-                                                             icon="x-mark"
-                                                             wire:confirm="Вы действительно хотите удалить этот атрибут? Удаление произойдет со всех товаров"></flux:button>
+                                                @if($this->user()->can('update-items'))
+                                                    <flux:button wire:click="deleteAttribute({{$mainAttribute}})"
+                                                                 variant="danger"
+                                                                 icon="x-mark"
+                                                                 wire:confirm="Вы действительно хотите удалить этот атрибут? Удаление произойдет со всех товаров"></flux:button>
+                                                @endif
+
                                             </div>
                                             @break
                                         @case('textarea')
                                             <div class="flex justify-between items-end">
                                                 <flux:textarea wire:model="form.attributes.{{$mainAttribute->id}}"
                                                                label="{{$mainAttribute->name}}"/>
-                                                <flux:button wire:click="deleteAttribute({{$mainAttribute}})" variant="danger"
-                                                             icon="x-mark"
-                                                             wire:confirm="Вы действительно хотите удалить этот атрибут? Удаление произойдет со всех товаров"></flux:button>
+                                                @if($this->user()->can('update-items'))
+                                                    <flux:button wire:click="deleteAttribute({{$mainAttribute}})"
+                                                                 variant="danger"
+                                                                 icon="x-mark"
+                                                                 wire:confirm="Вы действительно хотите удалить этот атрибут? Удаление произойдет со всех товаров"></flux:button>
+                                                @endif
+
                                             </div>
                                             @break
                                         @default
                                             <flux:input.group class="items-end">
                                                 <flux:input wire:model="form.attributes.{{$mainAttribute->id}}"
-                                                            label="{{$mainAttribute->name}}" type="{{$mainAttribute->type}}"/>
+                                                            label="{{$mainAttribute->name}}"
+                                                            type="{{$mainAttribute->type}}"/>
 
-                                                <flux:button wire:click="deleteAttribute({{$mainAttribute}})" variant="danger"
-                                                             icon="x-mark"
-                                                             wire:confirm="Вы действительно хотите удалить этот атрибут? Удаление произойдет со всех товаров"></flux:button>
+                                                @if($this->user()->can('update-items'))
+                                                    <flux:button wire:click="deleteAttribute({{$mainAttribute}})"
+                                                                 variant="danger"
+                                                                 icon="x-mark"
+                                                                 wire:confirm="Вы действительно хотите удалить этот атрибут? Удаление произойдет со всех товаров"></flux:button>
+                                                @endif
+
                                             </flux:input.group>
                                             @break
                                     @endswitch
@@ -74,36 +97,49 @@
                                     @switch($attribute->attribute->type)
                                         @case('boolean')
                                             <div class="flex justify-between items-center">
-                                                <flux:switch wire:model="form.attributes.{{$attribute->item_attribute_id}}"
-                                                             :label="$attribute->attribute->name"/>
-                                                <flux:button wire:click="deleteAttribute({{$attribute->attribute}})"
-                                                             variant="danger" icon="x-mark"
-                                                             wire:confirm="Вы действительно хотите удалить этот атрибут? Удаление произойдет со всех товаров"></flux:button>
+                                                <flux:switch
+                                                        wire:model="form.attributes.{{$attribute->item_attribute_id}}"
+                                                        :label="$attribute->attribute->name"/>
+                                                @if($this->user()->can('update-items'))
+                                                    <flux:button wire:click="deleteAttribute({{$attribute->attribute}})"
+                                                                 variant="danger" icon="x-mark"
+                                                                 wire:confirm="Вы действительно хотите удалить этот атрибут? Удаление произойдет со всех товаров"></flux:button>
+                                                @endif
+
                                             </div>
                                             @break
                                         @case('textarea')
                                             <div class="flex justify-between items-end">
-                                                <flux:textarea wire:model="form.attributes.{{$attribute->item_attribute_id}}"
-                                                               label="{{$attribute->attribute->name}}"/>
-                                                <flux:button wire:click="deleteAttribute({{$attribute->attribute}})"
-                                                             variant="danger" icon="x-mark"
-                                                             wire:confirm="Вы действительно хотите удалить этот атрибут? Удаление произойдет со всех товаров"></flux:button>
+                                                <flux:textarea
+                                                        wire:model="form.attributes.{{$attribute->item_attribute_id}}"
+                                                        label="{{$attribute->attribute->name}}"/>
+                                                @if($this->user()->can('update-items'))
+                                                    <flux:button wire:click="deleteAttribute({{$attribute->attribute}})"
+                                                                 variant="danger" icon="x-mark"
+                                                                 wire:confirm="Вы действительно хотите удалить этот атрибут? Удаление произойдет со всех товаров"></flux:button>
+                                                @endif
+
                                             </div>
                                             @break
                                         @default
                                             <flux:input.group class="items-end">
-                                                <flux:input wire:model="form.attributes.{{$attribute->item_attribute_id}}"
-                                                            label="{{$attribute->attribute->name}}"
-                                                            type="{{$attribute->attribute->type}}"/>
+                                                <flux:input
+                                                        wire:model="form.attributes.{{$attribute->item_attribute_id}}"
+                                                        label="{{$attribute->attribute->name}}"
+                                                        type="{{$attribute->attribute->type}}"/>
+                                                @if($this->user()->can('update-items'))
+                                                    <flux:button variant="danger" icon="x-mark"
+                                                                 wire:confirm="Вы действительно хотите удалить этот атрибут? Удаление произойдет со всех товаров"
+                                                                 wire:click="deleteAttribute({{$attribute->attribute}})"></flux:button>
+                                                @endif
 
-                                                <flux:button variant="danger" icon="x-mark"
-                                                             wire:confirm="Вы действительно хотите удалить этот атрибут? Удаление произойдет со всех товаров"
-                                                             wire:click="deleteAttribute({{$attribute->attribute}})"></flux:button>
                                             </flux:input.group>
                                             @break
                                     @endswitch
                                 @endforeach
-                                <livewire:item.item-attribute-dialog-form/>
+                                @if($this->user()->can('update-items'))
+                                    <livewire:item.item-attribute-dialog-form/>
+                                @endif
                             </div>
                         </div>
                     </flux:card>
@@ -121,7 +157,7 @@
                                 <div class="flex items-center gap-4">
                                     <flux:subheading>Был обновлён:</flux:subheading>
                                     <flux:badge
-                                        :color="$item->updated ? 'lime' : 'red'">{{$item->updated ? 'Да' : 'Нет'}}</flux:badge>
+                                            :color="$item->updated ? 'lime' : 'red'">{{$item->updated ? 'Да' : 'Нет'}}</flux:badge>
                                 </div>
                             </div>
                             <div class="space-y-6">
@@ -149,7 +185,7 @@
                                 <div class="flex items-center gap-4">
                                     <flux:subheading>Статус:</flux:subheading>
                                     <flux:badge
-                                        :color="$item->fromPrice ? 'lime' : 'red'">{{$item->fromPrice ? $item->fromPrice->message : "Товар не найден"}}</flux:badge>
+                                            :color="$item->fromPrice ? 'lime' : 'red'">{{$item->fromPrice ? $item->fromPrice->message : "Товар не найден"}}</flux:badge>
                                 </div>
                                 @if($item->fromPrice)
                                     <flux:input label="Артикул" :value="$item->fromPrice->article" readonly/>
